backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.4.2
email-validator==2.1.0

# HTTP client (for any external API calls)
httpx==0.25.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1